+++ changelog.md
@@ -6,9 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+- Added new API to Lock for seeing if a lock is read/write locked
 
 ## [1.1.0] - 2018-09-28
-### Added
+### Changed
 - Documentation fixes
 - Use go 1.11 modules for versioning
 - New feature:  CAR cache!